+++ siye-api/src/main/java/mx/com/teclo/siye/persistencia/hibernate/dao/catalogo/TipoPersonaDAO.java
@@ -5,5 +5,5 @@ import mx.com.teclo.siye.persistencia.hibernate.dto.catalogo.TipoPersonaDTO;
 
 public interface TipoPersonaDAO extends BaseDao<TipoPersonaDTO>{
 	
-	public TipoPersonaDTO getTipoPersonaXID(Integer idTipoPersona);
+	public TipoPersonaDTO getTipoPersonaXID(Long idTipoPersona);
 }
